Rename op_tracker to execution_timer
